fix: move ffi outside of own folder
