+++ src/optionals/optionals.h
@@ -29,7 +29,7 @@
 #include "object/object.h"
 #include "buffer.h"
 #include "unittest/unittest.h"
-#include "ffi/ffi.h"
+#include "ffi.h"
 
 typedef Value (*BuiltinModule)(DictuVM *vm);
